Sync user minutes online to AirdropBot and use for airdrop eligibility (#7563)

+++ backend/canisters/online_users/api/src/queries/online_minutes.rs
@@ -12,5 +12,5 @@ pub struct Args {
 #[ts_export(online_users, online_minutes)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
-    Success(u32),
+    Success(u16),
 }
